fix and edit buttons for MainActivity
This could be put up to debate, but I like the names "Save" and "Browse" better for the button displays. They match our activity names better and keep things consistent.

Also fixed an error for the saveNewEntry function, so that it can be used for the save button in the main xml file.