implement pin to log into app
* pin is checked before logging in
When the pin the user enters matches the pin in
sharedPreferences, then the user can log into the
app. The app goes to the MainActivity once the
password matches the pin in sharedPreferences

* start app in login activity
Here manifest is changed so that the app starts in the
Login activity. This is done by making sure the intent-filter
is in the login activity tag.

* edit debugging messages to use TAGs
Tags are found at the top of the class
Also some intelliJ fixes implemented

* get pin string from sharedPreferences
To get a value from sharedPreferences, you need to get it via
a key. A key for the pin was created in the PreLoginActivity.
All values in SharedPreferences are stored via key/values.

Co-authored-by: DES21-BYUI <des21@byui.edu>
Co-authored-by: Kira <56455045+kiracherie@users.noreply.github.com>