implementing RecyclerView with Firebase
A new class! DatabaseFacade which does all
the dirty work with the database. All you
need to know is that DatabaseFacade has two
main functions

addValue: which adds a new Entry to the database
getFirebaseRecyclerAdapter: which allows you to create an adapter for the database
which displays all the entries that the DatabaseFacade's
author has